requirements.txt
mtcnn
tensorflow==2.7.0
opencv-python
keras
scikit-learn
numpy
pillow
